DevelopmentSettings: Remove and force-disable automatic OTA
With this commit we are removing the user switch, but this is a System API and as Google
says in the original commit this value is set internally (and its code is within Google services too).
Indeed the related frameworks base commit just publishes the String, but the main code is
somewhere else.
So, to be sure the automatic update function is really kept disabled, we are forcing it to disabled
(it means we are enabling the "disable automatic ota" feature) at least once in the onCreate method.

Ticket: NOUGAT-29

Change-Id: I9b9d65795f693cbfddda630120f49c6446ba1b27
Signed-off-by: Alex Naidis <alex.naidis@linux.com>